Added HTTPS support and cert gen feature
Server can now run in HTTPS mode, which is enabled in the config.json.
When HTTPS is enabled it runs on the https_port and a companion HTTP
server is started on the http_port which simply redirects to the HTTPS
port address.

The new `npm start --gen-cert` | `node index.js --gen-cert` feature
is now added. This runs through generating a self-signed HTTPS
certificate that can be used with the server. The certificate and key
are set in the config.js.